Fixes a crash in ChromeTableViewControllerTest.
CheckSwitchCellStateAndText() EXPECTs that the given `switch_item`
responds to `isOn`, but if that expectation is not met, the helper
continues on to call the missing method, which leads to a crash. This
CL replaces the EXPECT with an ASSERT, returning early and avoiding
the crash.

BUG=None

Change-Id: I5c9c52db8aa663b963a61ee38fa72c1a8dfb1a55
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3690093
Reviewed-by: Justin Cohen <justincohen@chromium.org>
Commit-Queue: Rohit Rao <rohitrao@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1034011}
NOKEYCHECK=True
GitOrigin-RevId: 1adf8fa03b33d0edcc21530215192a0f0d35488e