[iOS] Add read more button to unfold URL in Incognito Interstitial
A new button is added in the Incognito interstitial so as to let the
user unfold the URL that is to be opened, rather than display it in its
entirety by default. It is now limited to 3 lines, and then only when
the user taps the "more" button this limit is removed. In order to have
more control over how the URL is displayed, the `subtitleText` of the
`PromoStyleViewController` is left empty and the Incognito interstitial
view controller has its own `urlText` property which is displayed in a
custom label located in the `PromoStyleViewController`'s specific
content view. A gradient view is also added on the leading end of the
new button.

Also the `PromoStyleViewController` is fixed to ensure the banner image
size is re-computed when some related properties are set.

Change-Id: I757ef0094eb9ebba7ec604b051c6d905f5dadfa6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3825909
Commit-Queue: Quentin Pubert <qpubert@google.com>
Reviewed-by: Mark Cogan <marq@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1035484}
NOKEYCHECK=True
GitOrigin-RevId: e4945d8df1abe3d435902a527b68a80172bdeea3