Move model provider factory to embedder to share with iOS
The model provider factory has code that is common for iOS and other
platforms. This object depends on optimization_guide and cannot be
inside the platform internals code. So, move it to embedder dir where
this code can be shared.
Ensure the default models are passed the same way in chrome and ios.

BUG=1346389

Change-Id: I3d12297f92b09635c2c7166fad9e05a9fcd5d5aa
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3830821
Commit-Queue: Siddhartha S <ssid@chromium.org>
Reviewed-by: Min Qin <qinmin@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1035609}
NOKEYCHECK=True
GitOrigin-RevId: 3ccaa89a7b3b556bdbbede3607f4bf7c26fadcf3