[iOS] Speculative fix for TabGrid indicator
It seems that the ActivityIndicator of the tab grid is shown and started
but is only hidden.
Try to also stop it to see if it reduces the flakes linked to the
indicator still executed.

Bug: 1352970
Change-Id: I62d36e527135397fcc186eb310ed446da62d7408
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3831543
Reviewed-by: Jérôme Lebel <jlebel@chromium.org>
Auto-Submit: Gauthier Ambard <gambard@chromium.org>
Commit-Queue: Jérôme Lebel <jlebel@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1036557}
NOKEYCHECK=True
GitOrigin-RevId: 4d2ae5cd8aa0369264bd4a8ed9283a715b31036c