[ios] Prevents double refreshing the Feed when opening an NTP
Both setWebstate and didLoadPageWithSuccess are called whenever an NTP is opened (either a new one, or returning to a tab with an NTP already in it). Because of this we would try to trigger a feed reload twice, which could lead to issues.

This CL fixes that by only allowing didLoadPageWithSuccess to attempt a feedRefresh.

Change-Id: I4f13c87c224314187b9c9c0343b829c5ee56b27a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3843232
Commit-Queue: Sergio Collazos <sczs@chromium.org>
Reviewed-by: edchin <edchin@google.com>
Cr-Commit-Position: refs/heads/main@{#1037352}
NOKEYCHECK=True
GitOrigin-RevId: bf08b063999ffdf8939a55f05e1e2729e95b0959