Disable AccountsTableTestCase potentially causing flake
One of the AccountsTableTestCase tests are causing the others to be
flaky. (The failure happens at the start of the test and appears as a
GridCell animation failure.)

This problem can be seen in
https://chrome-swarming.appspot.com/task?id=5cddd8c00b76a011&w=true
where the only two tests ran before the flake were
testDismissSignOutConfirmationTwice and
testInterruptDuringSignOutConfirmation. While those tests passed
successfully, it is believed one of them is causing flake.

The tests are disabled in this CL to verify if the flake clears up.

Bug: 1352968
Change-Id: If49870bb0915bde12158e8313a4566218cb9fa92
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3846292
Auto-Submit: Mike Dougherty <michaeldo@chromium.org>
Owners-Override: edchin <edchin@google.com>
Reviewed-by: edchin <edchin@google.com>
Commit-Queue: edchin <edchin@google.com>
Cr-Commit-Position: refs/heads/main@{#1037993}
NOKEYCHECK=True
GitOrigin-RevId: 7680b4c66f7a20094746131f1661cff6a143186b