Implement LeastRecentlyShown().
LeastRecentlyShown() enables the Promos Manager to select the least
recently shown promo given a set of active promos and an impression
history list.

This functionality is central to the Promos Manager's ability to select
the next promo for display when the manager's SceneAgent triggers a
promo for display.

Change-Id: Ib1728e4a7a787ebc0ec7e17adc983aa1f5ea0e6c
Bug: 1355288
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3842640
Reviewed-by: Ali Juma <ajuma@chromium.org>
Commit-Queue: Benjamin Williams <bwwilliams@google.com>
Cr-Commit-Position: refs/heads/main@{#1038040}
NOKEYCHECK=True
GitOrigin-RevId: b893d8f7f12fec043304bfbe2987c02d84d1b438