ios: Fix crash when adding bookmark from incognito.
The on-the-record BookmarkIneractionController is always used when
adding or editing bookmarks. A refactor in crrev.com/c/3755541 changed
the logic to use the on-the-record WebState (which can be null) when
determining the title and URL for a bookmark.

Instead, pass a BookmarkAddCommand with the correct WebState.

Fixed: 1353114
Change-Id: Iadf8de4445ad24e059d1889f5ac503d2887f81ae
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3846908
Auto-Submit: Justin Cohen <justincohen@chromium.org>
Commit-Queue: Rohit Rao <rohitrao@chromium.org>
Commit-Queue: Justin Cohen <justincohen@chromium.org>
Reviewed-by: Rohit Rao <rohitrao@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1038084}
NOKEYCHECK=True
GitOrigin-RevId: f69798c6b4b87084eb9a894b7a8f7be7e237ae5b