[iOS] Fix horizontal margins in Incognito interstitial
A new `titleHorizontalMargin` property is added to Promo style view
controller to enable customisation of margins around the title label.
If not set before view loads, it defaults to `kTitleHorizontalMargin`.

A new `stackViewHorizontalMargin` argument is added to the designated
initialisers of revamped and non-revamped Incognito views to customise
margins of the Incognito views as well.

Incognito view bouncing is disabled in Incognito interstitial as well
since this scroll view is not even supposed to be scrollable (it is
constrained to have the same height as its content).

Bug: 1355292
Change-Id: I3656cdacca4be8cc43a9f0a12d998d4f9c06874c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3845534
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Reviewed-by: Sylvain Defresne <sdefresne@chromium.org>
Commit-Queue: Quentin Pubert <qpubert@google.com>
Cr-Commit-Position: refs/heads/main@{#1038158}
NOKEYCHECK=True
GitOrigin-RevId: 18cd43b3534f8d8a82c8a8d044ae91bf3a353452