Moved some PasswordController tests to SharedPasswordController tests.
Some PasswordController tests only tested flows related to SharedPasswordController. These tests did not load any HTML, so they can live among SharedPasswordController's tests.

The reason behind this change is: Because no HTML was loaded, the frames used in this tests were null pointers. The tests worked so far like that, but crrev.com/c/3820103 introduces some changes which require those frame to exists.
The solution for that would be creating FakeWebFrame's, which need ot be tied to a FakeWebState. But PasswordController unit tests use a real WebState. Thus, the best option was for these tests to be moved alltogether to SharedPasswordController's tests, where it also makes more sense for them to live.

Bug: 1344776
Change-Id: I0b59a39109a8c9b91fed8b9d84f4a7f4f3601f54
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3854597
Reviewed-by: Maria Kazinova <kazinova@google.com>
Commit-Queue: Theodor-Stefan Cristea <theocristea@google.com>
Cr-Commit-Position: refs/heads/main@{#1038795}
NOKEYCHECK=True
GitOrigin-RevId: b4fca2a178f50251573ce0f13317c27f9e74b942