[iOS] Fix incognito NTP animation bug in a better way
In an initial patchset of crrev.com/c/3814153, this solution was
proposed, but as it's a bit more complicated, the simpler solution was
chosen, so as not to introduce any bugs in the cherry-pick. Now, the
better solution can be tested over a normal release cycle.

This removes the change from the above CL and adds a better solution.

Bug: 1347444
Change-Id: I6bfc46e7a9949802b84c174597ee099980eb0369
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3864430
Reviewed-by: Rohit Rao <rohitrao@chromium.org>
Commit-Queue: Rohit Rao <rohitrao@chromium.org>
Auto-Submit: Robbie Gibson <rkgibson@google.com>
Cr-Commit-Position: refs/heads/main@{#1041811}
NOKEYCHECK=True
GitOrigin-RevId: 1fce4748663916c32a0f5b5995dc609402daeccf