[iOS] Don't close Find In Page from new tab toolbar button
This button used to be the search button, and there, it made more sense
to close Find In Page when pressing it. Now, the button opens a new tab,
so that is unnecessary.

This CL also renames the method so it's clearer what the action is
doing.

Fixed: 1356994
Change-Id: I59dea3379abc3f58083ca811088db136393ba60f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3867626
Auto-Submit: Robbie Gibson <rkgibson@google.com>
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Commit-Queue: Gauthier Ambard <gambard@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1042078}
NOKEYCHECK=True
GitOrigin-RevId: 45fb6523ba68fe36c08ab17154fe8150dd2bd1cb