[iOS] Create Mock Infobar Delegate for Tailored Security Service
Create a mock infobar delegate for tailored security service. This CL
ony has the file initializations and base functionality of a mock
delegate. More functionality and usage will be added in follow-up
CLs so that CLs are easier to read and manage in general.

Bug: 1358531
Change-Id: I5f121f6a380cab0c36abef0cbd069b3bb6be4e8b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3866084
Reviewed-by: Ali Juma <ajuma@chromium.org>
Commit-Queue: Joemer Ramos <joemerramos@chromium.org>
Reviewed-by: Chris Lu <thegreenfrog@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1042120}
NOKEYCHECK=True
GitOrigin-RevId: 45cf3f21dd5035285c3bca293e92a1013c012408