Create RegisterForContinuousDisplay() & RegisterForSingleDisplay().
The PromosManager needs to expose two public methods enabling promos to
be registered for single or continuous display.

Although the active promos are stored in a list Pref, the order promo
entries are populated in the list is irrelevant, as active promos are
stored in an in-memory set during PromosManager evaluation.

A given promo can be registered both for single & continuous display.
The manager will handle this use-case gracefully.

Change-Id: I8e62d2995f776a8390ba4390912c92259a566d57
Bug: 1359359
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3864316
Reviewed-by: Robbie Gibson <rkgibson@google.com>
Commit-Queue: Benjamin Williams <bwwilliams@google.com>
Cr-Commit-Position: refs/heads/main@{#1042716}
NOKEYCHECK=True
GitOrigin-RevId: dabc7b90775b82d8d16c08b005282daa72cce7be