[ios] FRE client side trial update
In version 2 of MICeAndDefaultBrowserPromo trial, each of the three
default browser promo trials are pushed to 25% on all channels, which
leaves the default group to also have 25% users. MICe related arms are
removed. This CL is intended to be cherry-picked into M106.

Bug: 1359413
Change-Id: I6ebaa42464cfbb0e2b75e98636bd10945f5677bc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3870641
Commit-Queue: Ginny Huang <ginnyhuang@chromium.org>
Reviewed-by: Rohit Rao <rohitrao@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1042791}
NOKEYCHECK=True
GitOrigin-RevId: d985f81d6064cb7fbd9c55e899366d0dacc6799c