[ios] Disable kFastApplicationWillTerminate in developer build
The kFastApplicationWillTerminate experiment make it difficult to
debug crashes during application shutdown (as the developer may be
in the "enabled" pool).

Disable the experiment for Debug developer build to ensure that
developer will not be in the "enabled" pool and can reproduce the
application shutdown crashes that they are assigned to debug.

This would have made https://crbug.com/1351660 easier to debug
(as due to random, developer were all put in the "enabled" pool
and could not reproduce the issue localy when it is quite easy
to reproduce if the experiment is disabled).

Bug: 1328891
Change-Id: I395f7e192e82cfb6e878895c1f34e6d32e2de9f6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3870565
Reviewed-by: Justin Cohen <justincohen@chromium.org>
Commit-Queue: Sylvain Defresne <sdefresne@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1043154}
NOKEYCHECK=True
GitOrigin-RevId: f9812c88a3048d576373bd13e917cd0ee64df3d8