Update the return type of GetValues() of PolicyValueProvider to Dict
Return a dictionary that contains policy values from
PolicyValueProviders. Each value provider will return the policy values
in a dictionary with the corresponding ID. Returning dictionary will
make it easier to merge the values from individual providers easier and
independent from each other. The policy values are stored as dictionary
in the list already.
Send a list of policy IDs to Javascript to indicate the specific order
the UI should display the policies as Javascript doesn't store the
entries in a dictionary in the order they're added and we want the UI to
display the policies in a specific order.

Bug: b:233209041
Change-Id: I56c43c959c370c8e3cf36fd8a5278bd955ccc2c6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3780813
Reviewed-by: Rebekah Potter <rbpotter@chromium.org>
Reviewed-by: Owen Min <zmin@chromium.org>
Reviewed-by: Mike Dougherty <michaeldo@chromium.org>
Reviewed-by: Pavol Marko <pmarko@chromium.org>
Commit-Queue: Irem Uguz <iremuguz@google.com>
Cr-Commit-Position: refs/heads/main@{#1043181}
NOKEYCHECK=True
GitOrigin-RevId: a34ffd0841ba865f7df160593f83b5559cf4b041