[ios] Only Register client-side study for TrendingQueries for new users
A server-side finch config will capture all existing users before the
experiment to allow for easy ramping up, preperiod analysis, and
cross-talk analysis. The client-side study configuration is now
guarded with a check to only register if it is a new user. This can
mean a brand new user or a user that is returning that was a new
user during the experiment runtime. The study name is updated to
reflect that.

Bug: 1359488
Change-Id: I2f79f3d6734f027c82ee590f26c28d9a4d0a32a3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3866459
Reviewed-by: Robert Kaplow <rkaplow@chromium.org>
Reviewed-by: Ginny Huang <ginnyhuang@chromium.org>
Reviewed-by: Rohit Rao <rohitrao@chromium.org>
Commit-Queue: Chris Lu <thegreenfrog@chromium.org>
Reviewed-by: Sebastien Seguin-Gagnon <sebsg@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1044020}
NOKEYCHECK=True
GitOrigin-RevId: f7f3a0911161cd6167fe4eaff6b7467ab8a80ab4