[iOS] Fix the logic to delete session sections in RecentTabs
This CL changes the logic used to delete the session sections in Recent
Tabs. Previously, if the a non-last section was deleted, the section
identifiers were not consecutive after kFirstSessionSectionIdentifier.
As the previous logic was assuming this, it was creating issues where
the sections were not properly deleted, leading to crashes as the UI
was no longer representing the model.

Fixed: 1356603
Change-Id: Ia6edebee53db431ccef364fa1ea529dc2d739a77
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3870448
Auto-Submit: Gauthier Ambard <gambard@chromium.org>
Commit-Queue: Sergio Collazos <sczs@chromium.org>
Reviewed-by: Sergio Collazos <sczs@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1044182}
NOKEYCHECK=True
GitOrigin-RevId: 13f1df7fa0698cca1bd09d61d67d3059e645fff7