[iOS] Only pass image and not imageName for OverflowMenuItem
The imageName was added to provide a solution for changing Dark/Light
mode while the menu was presented (crbug.com/1324516).
As the fix isn't working for all the icons, the bug is fixed in iOS 16
and it is only happening when the overflow menu is on the screen
while the mode is changed, I think we can remove it.

Bug: 1315544
Change-Id: If7f21af75bb82e1bb596df4fed7436ccd112b69b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3880074
Commit-Queue: Robbie Gibson <rkgibson@google.com>
Reviewed-by: Robbie Gibson <rkgibson@google.com>
Auto-Submit: Gauthier Ambard <gambard@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1044562}
NOKEYCHECK=True
GitOrigin-RevId: fc62a17ac161fa1e039e451712cf505723e32b7a