Implement displayPromo.
This change introduces a basic implementation for displayPromo, which
grabs the corresponding feature-team handler or provider, and calls its
respective display method.

In follow-up CLs, displayPromo will grow to record the impression in the
model layer, as well as configure the handler or providers delegates.

Change-Id: Ib5b8353583bfada29989268311fc45920fab1f74
Bug: 1361392
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3882893
Commit-Queue: Benjamin Williams <bwwilliams@google.com>
Reviewed-by: Ali Juma <ajuma@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1044681}
NOKEYCHECK=True
GitOrigin-RevId: 67fd861cb60334b3b075552902a89cb063e0b886