[iOS] Store the creation time of each WebState
This CL stores the creation time of each WebState. The existing
WebStates will be treated as inifitely old.
For now this is not recorded as the metrics would be skewed as
all WebStates would be treated as inifitely old. It will be recorded
in a future release.

Bug: 1358074
Change-Id: Ie737de965d7544da6ba76485b6cf4e5ee775d05b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3876422
Reviewed-by: Sylvain Defresne <sdefresne@chromium.org>
Commit-Queue: Gauthier Ambard <gambard@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1045115}
NOKEYCHECK=True
GitOrigin-RevId: 1aebed36f7a680417b8df050c192568359fba2eb