[iOS] When reloading an error page, use the real URL for UserAgent
Before this CL, when the user was reloading an error page, the URL of
the error page was used to check if the page needed a User Agent or not.
This CL uses the URL of the failed navigation instead, which makes more
sense.

Fixed: 1360567
Change-Id: Iaa8f5cb81c11e719edabf41fc4d8a206b4ceaf88
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3883944
Reviewed-by: Ali Juma <ajuma@chromium.org>
Commit-Queue: Ali Juma <ajuma@chromium.org>
Auto-Submit: Gauthier Ambard <gambard@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1045135}
NOKEYCHECK=True
GitOrigin-RevId: 98e0f2e75bb229f59737cb8d943767cdfa1cbb7f