[PWM UI Split] Add "Export Passwords" button to submenu
This CL adds the "Export Passwords" button to the new Settings submenu.
I'm sorry for the size of this! It turned out that getting this button
right had a lot of dependencies (e.g., the mediator) and none of them
really made sense to do on their own.

The code in the ViewController is largely copied from the parent
directory's ViewController with a few minimal changes. One exception is
that the PasswordExporter now lives in the Mediator, rather than the
ViewController; this is IMO more consistent with the
VC/Coordinator/Mediator architecture patterns. This is mostly a
reorganization and doesn't come with any behavior changes to how the
button/export flow works.

A follow-up CL will remove this button from the Password Manager, and
at the same time update the EG tests to verify this implementation
instead.

Bug: 1335156
Change-Id: Iff8a462d4bcd7c314d9aded43c2b84baee49cb62
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3864253
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Commit-Queue: Tommy Martino <tmartino@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1045144}
NOKEYCHECK=True
GitOrigin-RevId: 8cf120d8492f7df0a83c7e82db15ea2480af54b6