[ios] Change ChromeIdentityService API to accept SystemIdentity
Duplicate all methods of ChromeIdentityService taking a ChromeIdentity*
to have another variant taking an id<SystemIdentity> and implement the
old methods by calling the new ones.

As the internal sub-classes overrides the old methods, this should not
cause any change to behaviour, yet it will allow to convert the internal
class to use the new API in a single CL (by removing the old overrides
as the forwarding methods in the super class will kick in to forward to
the newly overridden methods).

Some methods have to be duplicated (helper methods that invoke methods
from the class itself since if the conversion to id<SystemIdentity>
would be implemented there, there would not call the correct one).

This is part of a multi-CL change:
 => https://crrev.com/c/3874944
  - https://crrev.com/i/4965299
  - https://crrev.com/c/3874935

Bug: 1201182
Change-Id: I0004be8aa01ab3b2e8977f053d5d436c2a45651b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3874944
Reviewed-by: Jérôme Lebel <jlebel@chromium.org>
Commit-Queue: Sylvain Defresne <sdefresne@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1045816}
NOKEYCHECK=True
GitOrigin-RevId: 96bffb701f9e79f14cf5501d96b4df7231d16abe