IDNA 2008: Add UKM for deviation characters in hostnames
This CL records a UKM for deviation characters that appear in
hostnames when the hostname is typed in the omnibox and then
navigated to. This information will be used to understand what
hostnames are affected by crbug.com/694157.

Most of this CL is plumbing code to pass the deviation character bit
from the omnibox text entry down to the navigation code. The URL
data types (GURL, Origin etc) don't keep deviation characters in the
URL, so we can't simply use the navigated URL to record this metric.

New UKM collection review doc is at
https://docs.google.com/document/d/1XKVhUe5Kes8EUfVF3LB-K9mflrE44cVL5Cnf0JrDHoM/edit

Bug: 1351768
Change-Id: Ieb3631c5c638854cfe4ed8a9ec552e6c936121cc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3862865
Reviewed-by: Alexei Svitkine <asvitkine@chromium.org>
Reviewed-by: Dana Fried <dfried@chromium.org>
Reviewed-by: Justin Donnelly <jdonnelly@chromium.org>
Reviewed-by: Orin Jaworski <orinj@chromium.org>
Commit-Queue: Mustafa Emre Acer <meacer@chromium.org>
Reviewed-by: Chris Thompson <cthomp@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1046424}
NOKEYCHECK=True
GitOrigin-RevId: 342b287d86f3490dcbfe6306e882a808c89c4440