Add lens button to ntp omnibox.
The Lens button will always be displayed on the right of the voice
search button if Lens is enabled. It can be enabled if you have "Google"
as your default search provider and opt into #enable-lens-in-ntp.

This change does not move the hint text; that will be done in a separate
cl.

Bug: 1353430
Change-Id: I8b6a4bcba9d0d5b6a306e73335784df284bc5f44
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3868984
Reviewed-by: Sylvain Defresne <sdefresne@chromium.org>
Commit-Queue: Jason Hu <hujasonx@google.com>
Reviewed-by: Chris Lu <thegreenfrog@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1046521}
NOKEYCHECK=True
GitOrigin-RevId: 125c82e4ddda79491909caaf0f103329dc62ca4f