[ios] Always open a new tab when launching from a widget.
If the app is cold-started from a widget and the omnibox is focused in
an existing NTP, sometimes the NTP will end up in an incorrectly
scrolled state, leaving two omniboxes visible. We can work around this
bug by always forcing a new NTP to open, which seems to make it harder
for the bug to manifest.

Bug: 1363375
Change-Id: I578073864b9dfa77cf23a10aafd5a6de54615084
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3891045
Reviewed-by: Justin Cohen <justincohen@chromium.org>
Commit-Queue: Rohit Rao <rohitrao@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1046673}
NOKEYCHECK=True
GitOrigin-RevId: a2f680b4eafeb494b7537f1a96fa740613d13db7