[PWM UI Split] Remove Export button + add test
This CL removes the "Export Passwords..." button from the Password
Manager UI. This causes the EG tests to fail, so at the same time I am
updating them to instead test the export flow in the new submenu.

In order to disarm the reauthentication module in this submenu, and to
avoid adding new uses of TopPresentedViewController, this CL introduces
a new ScopedPasswordSettingsReauthModuleOverride class, which, while in
scope, causes all new PasswordSettingsCoordinators to use the provided
mock instead of the real implementation.

LOW_COVERAGE_REASON=it's complaining about a literal test class

Bug: 1335156
Change-Id: I5162b0722c4e78c32ee744975e9902372487cae8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3883736
Reviewed-by: Sebastien Seguin-Gagnon <sebsg@chromium.org>
Commit-Queue: Tommy Martino <tmartino@chromium.org>
Reviewed-by: Olivier Robin <olivierrobin@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1046899}
NOKEYCHECK=True
GitOrigin-RevId: 68ae15f6725bdc740b3524a46f19ed0bb298cd89