Fix app crash when scrolling follow management page
The crash happens when trying to get spec from an invalid url.

The root cause is that some followed website has a null web page url
but valid rss url. ex. macworld.com. So, to ensure we can fetch the
favicon, we check both webpage url and rss url, use the one that is
not nil. There's a ready-to-use property  URL  in
followedWebChannelItem that we can use to fetch the favicon.

This cl fixes the same issue that happens to first follow UI.

Bug: 1362161
Change-Id: I08c10252a1600ceaac528d4e0c3a9dbb0a430b6f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3889118
Reviewed-by: Sergio Collazos <sczs@chromium.org>
Commit-Queue: Tina Wang <tinazwang@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1047553}
NOKEYCHECK=True
GitOrigin-RevId: c23a7419a60e881da334efe902355628628a2816