[ios] Fix potential crash during app shutdown
The OmniboxMediator register itself as an Observer of the
TemplateURLService. As OmniboxController didn't cleanup
the registration in its -stop method, the registration was
only removed in OmniboxMediator -dealloc.

As the lifetime of the OmniboxMediator may be longer than
that of the OmniboxController (confirmed by the unittest
BrowserCoordinatorTest.ShareChromeApp where the object is
deallocated when the PlatformTest autoreleasepool) it is
possible for the object to try to unregister itself after
the TemplateURLService has been destroyed, leading to a
crash.

Fix the crash by explicitly resetting the reference to the
TemplateURLService in OmniboxMediator by setting it to null
in OmniboxController -stop and cleaning up the observer at
that point.

Bug: none
Change-Id: I07f6d7628d5b7354f8bb8b5621a4bbec56985c58
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3899102
Commit-Queue: Sylvain Defresne <sdefresne@chromium.org>
Auto-Submit: Sylvain Defresne <sdefresne@chromium.org>
Commit-Queue: Stepan Khapugin <stkhapugin@chromium.org>
Reviewed-by: Stepan Khapugin <stkhapugin@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1047958}
NOKEYCHECK=True
GitOrigin-RevId: 140c9557b7b563161acafe3140efc4dc02f48a76