Tentative fix for RL crash
There seems to be a crash when favicon loader synchronously return
attributed when loading the model.
Add the cell to the model before requesting the favicon to avoid any
issue.

Note: This is a tentative fix as I could not reproduce the issue.

Bug: 1246283
Change-Id: I1bebd7a32e903f28409d20c25f2be312d0e63393
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3902002
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Commit-Queue: Olivier Robin <olivierrobin@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1048584}
NOKEYCHECK=True
GitOrigin-RevId: f7f3ed4bf7e3882844bfba1a59d94a823fea8715