[iOS] Add file operation verification in egtest
EnterpriseLoadingScreenTestCase have been really flaky lately. The issue
was because the policy test server was returning a 410 HTTP error code,
which causes the app to crash. The crash is because the DM token was
deleted and initialized again when there is a 410 error, but the test
server was still sending a request too fast for the dm token to be
completely reinitialized, which causes the crash.

Why the test server was sending a 410 error? The enterprise loading
screen is shown only on fresh installation. To simulate this
environment, before and after each test all related files on the disk
are deleted. The problem was disk operation was not terminated that the
test started, so the flow was the following one: App started -> Start
removing files -> The App do not send registration request to the server
as there is already registration file on the disk -> Removing file end
-> Send a policy request to the server => Server send 410 error as the
device that sent the request is not registered.

This CL ensure that we are waiting for file operation to be finished.

Change-Id: I6c5a59cc2bd865cfbf813208e6ead97ffc9bb9ea
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3894292
Reviewed-by: Vincent Boisselle <vincb@google.com>
Commit-Queue: Aliona Dangla <alionadangla@chromium.org>
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1048597}
NOKEYCHECK=True
GitOrigin-RevId: 9d4563b1699e81a392c175958abbe3032b76836b