[PWM UI Split] Pipe "Offer to save..." pref thru mediator
With this change, the PasswordSettingsMediator will read whether the
"Offer to Save Passwords" pref is managed by policy and pass this to the
consumer. It will also pass the initial value (on/off) of this pref, and
observe for subsequent changes and forward these as well.

The EG test changes are mainly re-adding tests from
legacy_password_settings_egtest which were deleted from this file, with
minor modifications to test against the new toggle in the submenu.

LOW_COVERAGE_REASON=CL contains eg tests covering this code

Bug: 1335156
Change-Id: I5796f6c6cf06bdd9c38abd95b8c944298724471a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3900407
Reviewed-by: Viktor Semeniuk <vsemeniuk@google.com>
Commit-Queue: Tommy Martino <tmartino@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1048662}
NOKEYCHECK=True
GitOrigin-RevId: 93fe2b2feea98a7c88854672c168acacf721845c