[ios] Remove pre-render pref migration
The migration was introduced in October 2020 (via the CL
https://crrev.com/c/2426704). It should now be safe to
assume that all users have migrated and the code can be
removed.

Move the registration of kNetworkPredictionSetting to
browser_prefs.mm as it is simpler to register the pref.

Bug: 538573
Change-Id: I69fe0a6205e04b7a3ee5f2e5dc052438fb4d8161
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3890873
Auto-Submit: Sylvain Defresne <sdefresne@chromium.org>
Reviewed-by: Mark Cogan <marq@chromium.org>
Commit-Queue: Sylvain Defresne <sdefresne@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1049006}
NOKEYCHECK=True
GitOrigin-RevId: 9cf0f7f84cd7a19135d1f32497ac001efa90c520