breadcrumbs: make BreadcrumbManager a singleton
This change replaces the multiple BreadcrumbManagers (one per profile/
BrowserState, plus another in ApplicationBreadcrumbsLogger) with a
singleton instance.

This will enable a future change to make BreadcrumbManager the single
source of truth for the breadcrumb event log, rather than being
duplicated as it is now (in the BreadcrumbManagers,
CrashReporterBreadcrumbObserver, and BreadcrumbPersistentStorageManager
when it reads the log from file).

It also simplifies the observer logic in the breadcrumbs code, as the
BreadcrumbManagerObservers (CrashReporterBreadcrumbObserver and
BreadcrumbPersistentStorageManager) will observe the singleton when
created and stop observing it when destroyed.

A follow-up CL will make BreadcrumbManager get previous session
events, after which BreadcrumbPersistentStorageManager and
CrashReporterBreadcrumbObserver can get events directly from
BreadcrumbManager.

Change-Id: I853bccc6e8f98416b1a84cbfc52f82b48925b4aa
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3840405
Reviewed-by: Mike Dougherty <michaeldo@chromium.org>
Reviewed-by: Sylvain Defresne <sdefresne@chromium.org>
Commit-Queue: Jesse McKenna <jessemckenna@google.com>
Reviewed-by: Colin Blundell <blundell@chromium.org>
Reviewed-by: Dmitry Gozman <dgozman@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1049393}
NOKEYCHECK=True
GitOrigin-RevId: 558ca951c9f07c4a0035f3a77a662192f1a1af5b