Sync: Fix metrics recorded during profile startup
Some Sync-related metrics are recorded when the SyncServiceImpl gets
created, which happens very soon during the profile startup. The
problem is that there are various kinds of profiles for which Sync
doesn't make sense, such as guest profiles or system profiles (used for
the profile picker).

Long-term, the plan is to not even create a SyncService for these kinds
of profiles (see crbug.com/1284664).
Short-term, to fix the metrics as soon as possible, this CL introduces
"version 2"s of some affected histograms, which are only recorded for
"interesting" kinds of profiles. A side benefit of this approach is
that we'll be able to compare the old and new histograms, and thus see
how large the issue was.

Bug: 1362552,1284659
Change-Id: I14c4c5e47de3004c337f9c43ded4027e6ba3db81
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3899794
Commit-Queue: Marc Treib <treib@chromium.org>
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>
Reviewed-by: Mark Pearson <mpearson@chromium.org>
Reviewed-by: Mikel Astiz <mastiz@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1049506}
NOKEYCHECK=True
GitOrigin-RevId: a31ea038245368c92ad1a88da04b47f116349ee4