Skeleton of ContactInfoSyncBridge.
As we finally agreed to keep the name "ContactInfo", this CL adds a
skeleton implementation of the corresponding sync bridge.

Like the AutofillSyncBridge, the ContactInfoSyncBridge is owned by
WebDataServiceWrapper and observes AutofillWebDataService for changes
to AutofillProfiles. Currently no logic is implemented and enabling the
sync bridge is flag guarded.

Bug: 1348294
Change-Id: I84a5a5002d174e815ed4c75526020874b91ac3de
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3892396
Reviewed-by: Matthias Körber <koerber@google.com>
Commit-Queue: Florian Leimgruber <fleimgruber@google.com>
Reviewed-by: Mikel Astiz <mastiz@chromium.org>
Reviewed-by: Peter Kasting <pkasting@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1049759}
NOKEYCHECK=True
GitOrigin-RevId: 80a229a24931d46316b17ff9288d3abc0a39213e