[iOS] Fix initial RTL layout for New Overflow Menu and scroll tracking
The direct issue here was related to LazyHStack. The LazyHStack seems
to not interact well with the ScrollViewProxy, maybe because part of
the HStack isn't loaded at the right time. As the destination list
is a short, fixed size, the lazy-loading isn't really necessary. A
standard HStack can be used.

Testing this also uncovered a new issue, where the scroll tracking
wasn't working correctly for RTL languages, as the offset calculations
weren't taking RTL into account. This CL also fixes that.

It also refactors the code slightly to separate parts of the layout.

Bug: 1365282
Change-Id: I92ce6238ffe5f3a24f0c71912cafafccdf1dd440
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3905152
Reviewed-by: David Jean <djean@chromium.org>
Commit-Queue: Robbie Gibson <rkgibson@google.com>
Cr-Commit-Position: refs/heads/main@{#1049796}
NOKEYCHECK=True
GitOrigin-RevId: 321d82192236878b8cfb9e24eef9bb3b7f10cca4